move validation code to map widget
